retry_on_error implemented as a util
